LPD-16317 The test classes must be inside a package with the version, as in REST Builder that is the package pattern and it is used to know which is the version of each query/mutation
